requirements.txt
aiokafka==0.13.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
arabic-reshaper==3.0.0
asn1crypto==1.5.1
async-timeout==5.0.1
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
cssselect2==0.9.0
distro==1.9.0
ecdsa==0.19.1
Events==0.5
fastapi==0.135.1
freetype-py==2.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.69.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.1
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
kafka-python==2.3.0
kafka-python-ng==2.2.3
lxml==6.0.2
Markdown==3.10.2
neo4j==6.1.0
opensearch-protobufs==0.19.0
opensearch-py==3.1.0
oscrypto==1.3.0
packaging==26.0
passlib==1.7.4
pillow==12.1.1
playwright==1.58.0
proto-plus==1.27.2
protobuf==5.29.6
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycairo==1.29.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
pyHanko==0.34.1
pyhanko-certvalidator==0.30.2
pyparsing==3.3.2
pypdf==6.9.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
pytz==2026.1.post1
PyYAML==6.0.3
reportlab==4.4.10
requests==2.32.5
rlPyCairo==0.4.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.52.1
svglib==1.6.0
tenacity==9.1.4
tinycss2==1.5.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
uritools==6.0.1
urllib3==2.6.3
uvicorn==0.41.0
webencodings==0.5.1
websockets==16.0
xhtml2pdf==0.2.17
